refactor(logging): replace LOG_BLE with LOG_INFO
* Updated logging calls in various device connection methods to use LOG_INFO for consistency.
* Modified device model handling in NormalizedShotData to use a fixed-size character array to prevent dangling pointers.
* Adjusted MQTT connection logging to improve clarity and maintainability.

+++ ESP32-S3-firmware/include/MqttManager.h
@@ -59,7 +59,7 @@ public:
   // Event publishers - called by TimerApplication
   void publishConnectionState(DeviceConnectionState state, const char* deviceName, const char* deviceModel);
   void publishDeviceInfo(const char* deviceName, const char* deviceModel, const char* firmwareVersion);
-  void publishSessionStarted(uint32_t sessionId, uint16_t startDelaySeconds);
+  void publishSessionStarted(uint32_t sessionId, float startDelaySeconds);
   void publishSessionStopped(uint32_t sessionId, uint16_t totalShots, uint32_t lastShotTimeMs = 0);
   void publishSessionSuspended(uint32_t sessionId);
   void publishSessionResumed(uint32_t sessionId);
